extract systems and resources to modules
